Reorganize repo.
Renamed lib/ by networks/.
Moved everything related to networks-specific to networks/ directory.
Renamed lib/networks/types to models/.
