--- wikichron/lib/networks/types/__init__.py
@@ -1,1 +1,0 @@
-
