+++ wikichron/networks/controls_sidebar_decorator/BaseControlsSidebarDecorator.py
