+++ wikichron/networks/cytoscape_stylesheet/BaseStylesheet.py
