+++ wikichron/networks/models/BaseNetwork.py
